feature: permette di giocare i livelli

+++ app/src/main/java/it/unicam/cs/mpgc/rpg129852/dto/Npc.java
@@ -6,5 +6,5 @@ public record Npc(
         String name,
         String imagePath,
         ProblemType problemType,
-        int maxProblemValue
+        double maxProblemValue
 ) {}
